+++ units/CMakeLists.txt
@@ -1,8 +1,11 @@
 # SPDX-License-Identifier: GPL-3.0-or-later
 # Copyright (c) 2019 Scipp contributors (https://github.com/scipp)
-set(INC_FILES include/scipp/units/unit.h)
+set(INC_FILES
+    include/scipp/units/unit.h
+    include/scipp/units/unit.tcc
+    include/scipp/units/neutron.h)
 
-set(SRC_FILES unit.cpp)
+set(SRC_FILES neutron.cpp)
 
 add_library(scipp-units STATIC ${INC_FILES} ${SRC_FILES})
 target_link_libraries(scipp-units PUBLIC Boost::boost)
