+++ units/neutron.cpp
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+// Copyright (c) 2019 Scipp contributors (https://github.com/scipp)
+/// @file
+/// @author Simon Heybrock
+#include "scipp/units/neutron.h"
+#include "scipp/units/unit.tcc"
+
+namespace scipp::units {
+INSTANTIATE(neutron::supported_units);
+} // namespace scipp::units
